+++ iree/hal/local/task_queue.h
@@ -10,9 +10,9 @@
 #include <stdint.h>
 
 #include "iree/base/api.h"
+#include "iree/base/internal/arena.h"
 #include "iree/base/internal/synchronization.h"
 #include "iree/hal/api.h"
-#include "iree/hal/local/arena.h"
 #include "iree/hal/local/task_queue_state.h"
 #include "iree/task/executor.h"
 #include "iree/task/scope.h"
